msg/muli/sync: Change playerId to uint8_t

+++ Source/dvlnet/abstract_net.h
@@ -19,7 +19,7 @@ public:
 	virtual int create(std::string addrstr) = 0;
 	virtual int join(std::string addrstr) = 0;
 	virtual bool SNetReceiveMessage(uint8_t *sender, void **data, size_t *size) = 0;
-	virtual bool SNetSendMessage(int dest, void *data, size_t size) = 0;
+	virtual bool SNetSendMessage(uint8_t dest, void *data, size_t size) = 0;
 	virtual bool SNetReceiveTurns(char **data, size_t *size, uint32_t *status) = 0;
 	virtual bool SNetSendTurn(char *data, size_t size) = 0;
 	virtual void SNetGetProviderCaps(struct _SNETCAPS *caps) = 0;
