+++ Source/dvlnet/base.h
@@ -18,7 +18,7 @@ namespace net {
 class base : public abstract_net {
 public:
 	bool SNetReceiveMessage(uint8_t *sender, void **data, size_t *size) override;
-	bool SNetSendMessage(int playerId, void *data, size_t size) override;
+	bool SNetSendMessage(uint8_t playerId, void *data, size_t size) override;
 	bool SNetReceiveTurns(char **data, size_t *size, uint32_t *status) override;
 	bool SNetSendTurn(char *data, size_t size) override;
 	void SNetGetProviderCaps(struct _SNETCAPS *caps) override;
